refine general preferences language menu
