+++ website/src/components/ui/ChatInput/ChatInput.module.css
@@ -157,74 +157,6 @@
   box-shadow: var(--ring-focus);
 }
 
-.language-select-wrapper {
-  position: relative;
-  display: flex;
-  align-items: center;
-  height: 100%;
-}
-
-.language-trigger {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  gap: var(--seg-arrow-gap, 12px);
-  height: 100%;
-  padding: 0;
-  border: none;
-  border-radius: calc(var(--seg-r, 22px) - 6px);
-  background: transparent;
-  color: inherit;
-  font: inherit;
-  letter-spacing: inherit;
-  text-transform: inherit;
-  cursor: pointer;
-  transition:
-    background-color 0.2s ease,
-    color 0.2s ease;
-}
-
-.language-trigger[data-open="true"],
-.language-trigger:hover {
-  /*
-   * 交互取舍：采用文字色微调强化状态感，同时保持透明背景以减少分段噪声。
-   * 可通过 --seg-hover-foreground 定制不同主题下的可访问对比度。
-   */
-  color: var(
-    --seg-hover-foreground,
-    color-mix(in srgb, var(--sb-text) 88%, white 12%)
-  );
-}
-
-.language-trigger:focus-visible {
-  outline: none;
-  box-shadow: var(--ring-focus);
-}
-
-.language-trigger-content {
-  display: inline-flex;
-  align-items: center;
-  gap: 0;
-}
-
-/*
- * 背景说明：语言缩写需要固定容积以保持箭头与动作按钮对齐。
- *  - 选用 4ch 结合 tabular-nums 提供稳定的等宽节奏，避免切换语言时触发抖动。
- *  - 暴露 --seg-code-width 便于未来适配更长徽标，无需改动组件结构。
- */
-.language-trigger-code {
-  display: inline-flex;
-  justify-content: center;
-  align-items: center;
-  inline-size: var(--seg-code-width, 4ch);
-  letter-spacing: normal;
-  font-variant-numeric: tabular-nums;
-}
-
-.language-trigger-label {
-  display: none;
-}
-
 .language-arrow,
 .language-swap {
   display: inline-flex;
@@ -392,112 +324,4 @@
   border: 0;
   opacity: 0;
   pointer-events: none;
-}
-
-.language-menu {
-  list-style: none;
-  margin: 0;
-  padding: var(--sb-menu-padding, 8px);
-  width: min(var(--sb-menu-width, 360px), calc(100vw - 32px));
-  max-height: var(--sb-menu-max-h, 420px);
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-  border-radius: var(--sb-menu-radius, 16px);
-  background: var(--menu-bg);
-  border: 1px solid var(--menu-border);
-  box-shadow: var(--sb-menu-shadow);
-  overflow-y: auto;
-  scrollbar-width: thin;
-  scrollbar-color: var(--menu-scroll-thumb) var(--menu-scroll-track);
-  opacity: 0;
-  transform: translateY(4px) scale(0.98);
-  transform-origin: top center;
-  transition:
-    opacity 0.2s ease,
-    transform 0.2s ease;
-}
-
-.language-menu[data-open="true"] {
-  opacity: 1;
-  transform: translateY(0) scale(1);
-}
-
-.language-menu::-webkit-scrollbar {
-  width: 8px;
-}
-
-.language-menu::-webkit-scrollbar-track {
-  background: var(--menu-scroll-track);
-  border-radius: 8px;
-}
-
-.language-menu::-webkit-scrollbar-thumb {
-  background: var(--menu-scroll-thumb);
-  border-radius: 8px;
-}
-
-.language-menu::-webkit-scrollbar-thumb:hover {
-  background: var(--menu-scroll-thumb-hover);
-}
-
-.language-menu-item {
-  margin: 0;
-  padding: 0;
-}
-
-.language-menu-button {
-  width: 100%;
-  display: grid;
-  grid-template-columns: 12ch 1fr auto;
-  align-items: center;
-  gap: 12px;
-  padding-inline: 16px;
-  height: 48px;
-  border: none;
-  border-radius: 12px;
-  background: transparent;
-  color: var(--menu-name-color);
-  font-size: var(--sb-font-sm, 15px);
-  font-weight: 500;
-  text-align: left;
-  cursor: pointer;
-  transition:
-    background-color 0.15s ease,
-    color 0.15s ease;
-}
-
-.language-menu-button:hover,
-.language-menu-button:focus-visible,
-.language-menu-button[data-active="true"] {
-  background: var(--menu-hover);
-  outline: none;
-}
-
-.language-option-code {
-  color: var(--menu-code-color);
-  font-weight: 700;
-  letter-spacing: var(--sb-code-letter-spacing, 0.08em);
-  text-transform: uppercase;
-}
-
-.language-option-name {
-  color: var(--menu-name-color);
-  font-weight: 500;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.language-option-check {
-  justify-self: end;
-  display: inline-flex;
-  align-items: center;
-  color: var(--menu-check-color);
-  opacity: 0;
-  transition: opacity 0.15s ease;
-}
-
-.language-menu-button[data-active="true"] .language-option-check {
-  opacity: 1;
 }
